add user-restaurant relationship migrations & implement auth check in restaurants feature

+++ routes/restaurants.js
@@ -2,13 +2,18 @@ const express = require('express')
 const router = express.Router()
 
 const db = require('../models')
+const restaurant = require('../models/restaurant')
 const Restaurant = db.Restaurant
 
 //Read
 router.get('/', (req, res, next) => {
   const keywords = req.query.keyword || req.query.category
-  return Restaurant.findAll(
-    {raw:true}
+  const userId =  req.user.id
+
+  return Restaurant.findAll({
+    where: { userId },
+    raw:true
+  }
   )
   .then((restaurants) => {
     return keywords ? restaurants.filter(rt => rt.name.includes(keywords) || rt.category.includes(keywords) ) : restaurants
@@ -23,10 +28,19 @@ router.get('/', (req, res, next) => {
 })
 router.get('/:id', (req, res, next) => {
   const id = req.params.id
+  const userId =req.user.id
   return Restaurant.findByPk(id,{
     raw:true
   })
   .then((restaurant) => {
+    if (!restaurant) {
+      req.flash('error', '找不到資料')
+      return res.redirect('/Restaurants')
+    }
+    if (restaurant.userId !== userId) {
+      req.flash('error', '權限不足')
+      return res.redirect('/Restaurants')
+    }
     res.render('show', {restaurant}) 
   })
   .catch((error) => {
@@ -41,6 +55,7 @@ router.get('/one/new', (req, res) => {
 })
 router.post('/', (req, res, next) => {
   const body =req.body
+  const userId = req.user.id
   return Restaurant.create({
     name: body.name,
     name_en: body.name_en,
@@ -50,7 +65,8 @@ router.post('/', (req, res, next) => {
     phone: body.phone,
     google_map: body.google_map,
     rating: body.rating,
-    description:body.description
+    description:body.description,
+    userId: userId
   })
   .then(() => {
     req.flash('success', '新增成功')
@@ -65,10 +81,19 @@ router.post('/', (req, res, next) => {
 //Update
 router.get('/:id/edit', (req, res, next) => {
   const id = req.params.id
+  const userId = req.user.id
   return Restaurant.findByPk(id,{
     raw:true
   })
   .then((restaurant) => {
+    if (!restaurant) {
+      req.flash('error', '找不到資料')
+      return res.redirect('/Restaurants')
+    }
+    if (restaurant.userId !== userId) {
+      req.flash('error', '權限不足')
+      return res.redirect('/Restaurants')
+    }    
     req.flash('success', '更新成功')
     return res.render('edit',{restaurant})
   })
@@ -80,39 +105,64 @@ router.get('/:id/edit', (req, res, next) => {
 router.put('/:id', (req, res, next) => {
   const body = req.body
   const id = req.params.id
-  return Restaurant.update({
-    name: body.name,
-    name_en: body.name_en,
-    category:body.category,
-    image:body.image,
-    location:body.location,
-    phone: body.phone,
-    google_map: body.google_map,
-    rating: body.rating,
-    description:body.description
-},{where:{id:id}})
-  .then(() => {
-    req.flash('success', '更新成功')
-    return res.redirect(`/Restaurants/${id}`)
-  })
-  .catch((error) => {
-    error.errorMessage = '更新失敗'
-    next(error)
-  })
+  const userId = req.user.id
+  return Restaurant.findByPk(id)
+    .then((restaurant) => {
+      if (!restaurant) {
+        req.flash('error', '找不到資料')
+        return res.redirect('/Restaurants')
+      }
+      if (restaurant.userId !== userId) {
+        req.flash('error', '權限不足')
+        return res.redirect('/Restaurants')        
+      }
+      return Restaurant.update({
+        name: body.name,
+        name_en: body.name_en,
+        category:body.category,
+        image:body.image,
+        location:body.location,
+        phone: body.phone,
+        google_map: body.google_map,
+        rating: body.rating,
+        description:body.description
+        },{where:{id:id}})
+          .then(() => {
+            req.flash('success', '更新成功')
+            return res.redirect(`/Restaurants/${id}`)
+           })
+        })
+    .catch((error) => {
+      error.errorMessage = '更新失敗'
+      next(error)
+    })
 })
 
 //Delete
 router.delete('/:id', (req, res, next) => {
   const id =req.params.id
-  return Restaurant.destroy({where:{id}})
-  .then(() => {
-    req.flash('success', '刪除成功')
-    res.redirect('/Restaurants')
-  })
-  .catch((error) => {
-    error.errorMessage = '刪除失敗'
-    next(error)    
-  })
+  const userId = req.user.id 
+
+  return Restaurant.findByPk(id)
+    .then((restaurant) => {
+      if (!restaurant) {
+        req.flash('error', '找不到資料')
+        return res.redirect('/Restaurants')
+      }
+      if (restaurant.userId !== userId) {
+        req.flash('error', '權限不足')
+        return res.redirect('/Restaurants')        
+      }
+      return Restaurant.destroy({where:{id}})
+      .then(() => {
+        req.flash('success', '刪除成功')
+        res.redirect('/Restaurants')
+      })
+    })
+    .catch((error) => {
+      error.errorMessage = '刪除失敗'
+      next(error)    
+    })
 })
 
 module.exports = router
